sre.wdqs: Integrate wcqs with wdqs cookbooks
Sets appropriate mutation topic, service names, etc when invoked
against wcqs instances.

Bug: T293638

+++ cookbooks/sre/wdqs/data-transfer.py
@@ -19,17 +19,27 @@ from time import sleep
 from spicerack.kafka import ConsumerDefinition
 from spicerack.remote import RemoteExecutionError
 
-from cookbooks.sre.wdqs import check_host_is_wdqs, wait_for_updater, get_site, get_hostname, MUTATION_TOPIC
+from cookbooks.sre.wdqs import check_hosts_are_valid, wait_for_updater, get_site, get_hostname, MUTATION_TOPICS
 
 BLAZEGRAPH_INSTANCES = {
     'categories': {
         'services': ['wdqs-categories'],
+        'data_path': '/srv/wdqs',
         'files': ['/srv/wdqs/categories.jnl', '/srv/wdqs/aliases.map'],
+        'valid_on': 'wdqs',
     },
-    'blazegraph': {
+    'wikidata': {
         'services': ['wdqs-updater', 'wdqs-blazegraph'],
+        'data_path': '/srv/wdqs',
         'files': ['/srv/wdqs/wikidata.jnl'],
-    }
+        'valid_on': 'wdqs',
+    },
+    'commons': {
+        'services': ['wcqs-updater', 'wcqs-blazegraph'],
+        'data_path': '/srv/query_service',
+        'files': ['/srv/query_service/wcqs.jnl'],
+        'valid_on': 'wcqs',
+    },
 }
 
 __title__ = "WDQS data transfer cookbook"
@@ -120,7 +130,7 @@ def run(args, spicerack):
     """Required by Spicerack API."""
     remote = spicerack.remote()
     remote_hosts = remote.query("{source},{dest}".format(source=args.source, dest=args.dest))
-    check_host_is_wdqs(remote_hosts, remote)
+    host_kind = check_hosts_are_valid(remote_hosts, remote)
 
     icinga_hosts = spicerack.icinga_hosts(remote_hosts.hosts)
     puppet = spicerack.puppet(remote_hosts)
@@ -138,8 +148,13 @@ def run(args, spicerack):
         raise ValueError("Only one destination node is needed. Not {total}({source})".
                          format(total=len(source), source=source))
 
-    services = BLAZEGRAPH_INSTANCES[args.blazegraph_instance]['services']
-    files = BLAZEGRAPH_INSTANCES[args.blazegraph_instance]['files']
+    instance = BLAZEGRAPH_INSTANCES[args.blazegraph_instance]
+    if host_kind != instance['valid_on']:
+        raise ValueError('Instance (valid_on:{}) is not valid for selected hosts ({})'.format(
+            instance['valid_on'], host_kind))
+
+    services = instance['services']
+    files = instance['files']
 
     stop_services_cmd = " && ".join(["systemctl stop " + service for service in services])
     services.reverse()
@@ -159,29 +174,32 @@ def run(args, spicerack):
                 _copy_file(source, dest, file)
                 dest.run_sync('chown blazegraph: "{file}"'.format(file=file))
 
-            if args.blazegraph_instance == 'blazegraph':
+            if args.blazegraph_instance in ('wikidata', 'commons'):
                 logger.info('Touching "data_loaded" file to show that data load is completed.')
-                dest.run_sync('touch /srv/wdqs/data_loaded')
+                dest.run_sync('touch {data_path}/data_loaded'.format(
+                    data_path=instance['data_path']))
 
             if args.blazegraph_instance == 'categories':
                 logger.info('Reloading nginx to load new categories mapping.')
                 dest.run_sync('systemctl reload nginx')
 
-            logger.info('Transferring Kafka offsets')
-            kafka = spicerack.kafka()
-
             source_hostname = get_hostname(args.source)
             dest_hostname = get_hostname(args.dest)
-            kafka.transfer_consumer_position([MUTATION_TOPIC],
-                                             ConsumerDefinition(get_site(source_hostname, spicerack), 'main',
-                                                                source_hostname),
-                                             ConsumerDefinition(get_site(dest_hostname, spicerack), 'main',
-                                                                dest_hostname))
+            if args.blazegraph_instance in MUTATION_TOPICS:
+                logger.info('Transferring Kafka offsets')
+                kafka = spicerack.kafka()
+                kafka.transfer_consumer_position([MUTATION_TOPICS[args.blazegraph_instance]],
+                                                 ConsumerDefinition(get_site(source_hostname, spicerack), 'main',
+                                                                    source_hostname),
+                                                 ConsumerDefinition(get_site(dest_hostname, spicerack), 'main',
+                                                                    dest_hostname))
 
             logger.info('Starting services [%s]', start_services_cmd)
             remote_hosts.run_sync(start_services_cmd)
-            wait_for_updater(prometheus, get_site(source_hostname, spicerack), source)
-            wait_for_updater(prometheus, get_site(dest_hostname, spicerack), dest)
+
+            if args.blazegraph_instance in MUTATION_TOPICS:
+                wait_for_updater(prometheus, get_site(source_hostname, spicerack), source)
+                wait_for_updater(prometheus, get_site(dest_hostname, spicerack), dest)
 
             if args.with_lvs:
                 logger.info('pooling %s', remote_hosts)
